requirements.txt
fastapi
uvicorn
python-dotenv
pinecone
openai
ollama
PyPDF2